[Chore/#20] Account - 글자수 반영 안되는 문제 해결
